# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handler.ErrorDetail:
    properties:
      code:
        example: INVALID_INPUT
        type: string
      message:
        example: name is required
        type: string
    type: object
  internal_handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_handler.ErrorDetail'
    type: object
...
  internal_handler.MessageData:
    properties:
      message:
        example: logged out successfully
        type: string
    type: object
  internal_handler.MessageResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handler.MessageData'
    type: object
...
host: localhost:8000
info:
  contact:
    name: FinanceOS Support
    url: https://github.com/financeos
  description: API completa para gestão financeira pessoal — contas, transações, orçamentos,
    investimentos, metas e IA.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: FinanceOS API
  version: 1.0.0
paths:
...
  /imports/csv:
    post:
      consumes:
      - multipart/form-data
      description: Importa transações de arquivo CSV com mapeamento de colunas (plano
        Pro)
      parameters:
      - description: Arquivo CSV
        in: formData
        name: file
        required: true
        type: file
      - description: UUID da conta destino
        in: formData
        name: account_id
        required: true
        type: string
      - description: JSON de mapeamento de colunas
        in: formData
        name: mapping
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
        "402":
          description: Plano insuficiente
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Importar extrato CSV
      tags:
      - Imports
  /imports/csv/preview:
    post:
      consumes:
      - multipart/form-data
      description: Retorna as primeiras linhas do CSV para conferência (plano Pro)
      parameters:
      - description: Arquivo CSV
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
        "402":
          description: Plano insuficiente
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Pré-visualizar CSV antes de importar
      tags:
      - Imports
  /imports/ofx:
    post:
      consumes:
      - multipart/form-data
      description: Importa transações a partir de arquivo OFX/QFX (plano Pro)
      parameters:
      - description: Arquivo OFX
        in: formData
        name: file
        required: true
        type: file
      - description: UUID da conta destino
        in: formData
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "402":
          description: Plano insuficiente
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Importar extrato OFX
      tags:
      - Imports
...
securityDefinitions:
  BearerAuth:
    description: 'Digite: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"